debugging commit
git-svn-id: svn://orca.physics.unc.edu/Orca@2440 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/HV/AmrelHV/ORAmrelHVModel.m
@@ -302,7 +302,7 @@ NSString* ORAmrelHVModelDataIsValidChanged	= @"ORAmrelHVModelDataIsValidChanged"
 	if(pollTime == 0 )return;
 	if([cmdQueue count] == 0){
 		[self getAllValues];
-		int nextPollTime = pollTime;
+		float nextPollTime = pollTime;
 		if(   rampState[0] != kAmrelHVNotRamping 
 		   || rampState[1] != kAmrelHVNotRamping) nextPollTime = .2;
